LayoutTests/imported/w3c/web-platform-tests/dom/events/scrolling/scrollend-event-not-fired-on-no-scroll-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: No scrollend event received for target [object HTMLDocument]
target
doc

Harness Error (FAIL), message = Unhandled rejection

PASS No scroll via wheel on div shouldn't fire scrollend.
TIMEOUT No scroll via keys on div shouldn't fire scrollend. Test timed out
NOTRUN No scroll via wheel on document shouldn't fire scrollend.
NOTRUN No scroll via keys on document shouldn't fire scrollend.
